Add CustomPost endpoint in FeedPage
